staging: wilc1000: remove unnecesary type cast of bss_type
There is no need to use type cast of bss_type because
hif_drv->cfg_values.bss_type is u8.

+++ drivers/staging/wilc1000/host_interface.c
@@ -434,7 +434,7 @@ static void handle_cfg_param(struct wilc_vif *vif,
 			wid_list[i].val = (s8 *)&bss_type;
 			wid_list[i].type = WID_CHAR;
 			wid_list[i].size = sizeof(char);
-			hif_drv->cfg_values.bss_type = (u8)bss_type;
+			hif_drv->cfg_values.bss_type = bss_type;
 		} else {
 			netdev_err(vif->ndev, "check value 6 over\n");
 			goto unlock;
